# Implement trait-based static upcast