--- a/examples/signals/src/lib.rs
+++ b/examples/signals/src/lib.rs
@@ -75,11 +75,10 @@ impl SignalSubscriber {
             .unwrap();
         let emitter = unsafe { emitter.assume_safe() };
 
-        let object = owner.cast().unwrap();
         emitter
             .connect(
                 GodotString::from_str("tick"),
-                object,
+                owner,
                 GodotString::from_str("notify"),
                 VariantArray::new_shared(),
                 0,
@@ -88,7 +87,7 @@ impl SignalSubscriber {
         emitter
             .connect(
                 GodotString::from_str("tick_with_data"),
-                object,
+                owner,
                 GodotString::from_str("notify_with_data"),
                 VariantArray::new_shared(),
                 0,