This replaces the old generated upcast methods on objects, and makes it so that dynamic casts can only be used for conversion to actual subclasses.

Unfortunately, standard conversion traits cannot be implemented due to conflict with blanket impls.

This did not have a negative impact on build times. The timings on my machine are as follows:

| | `dev` | `release` |
|-|-|-|
| This PR | 51.1s | 51.8s |
| `master` | 51.1s | 53.47s |

Close #252